backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
motor==3.6.0
httpx==0.27.2
python-dotenv==1.0.1
beautifulsoup4==4.12.3
pytest==8.3.3
anyio==4.6.2

